Move file cache match in lib/path/index
Compute digest whenever it is not computed to avoid uncached files.
