+++ lib/path/analyzers/file.js
@@ -1,9 +1,6 @@
 'use strict'
 
-const fs = require('fs')
-const hasha = require('hasha')
-
-async function analyzeFile(token, options) {
+async function analyzeFile(token) {
   if (token.analyzed) {
     return false
   }
@@ -12,19 +9,6 @@ async function analyzeFile(token, options) {
 
   if (!token.fileSize && token.stats) {
     token.fileSize = token.stats.size
-  }
-
-  if (!token.digest) {
-    const digest = await hasha.fromStream(fs.createReadStream(token.path), {
-      algorithm: options.digestAlgorithm,
-      encoding: 'base64'
-    })
-    token.digest = `${options.digestAlgorithm}-${digest}`
-
-    if (options.cache && options.cache.getFileCache && await options.cache.getFileCache(token, options)) {
-      token.unchanged = true
-      token.analyzed = true
-    }
   }
 
   token.analyzed = true
