+++ lib/path/index.js
@@ -2,6 +2,7 @@
 
 const fs = require('fs')
 const util = require('util')
+const hasha = require('hasha')
 
 const analyzeDirectory = require('./analyzers/directory')
 const analyzeName = require('./analyzers/name')
@@ -10,6 +11,14 @@ const analyzeArchive = require('./analyzers/archive')
 const analyzeFile = require('./analyzers/file')
 
 const lstat = util.promisify(fs.lstat)
+
+async function computeDigest(token, options) {
+  const digest = await hasha.fromStream(fs.createReadStream(token.path), {
+    algorithm: options.digestAlgorithm,
+    encoding: 'base64'
+  })
+  token.digest = `${options.digestAlgorithm}-${digest}`
+}
 
 async function analyzePath(token, options) {
   if (token.analyzed) {
@@ -26,6 +35,15 @@ async function analyzePath(token, options) {
     } else if (token.stats.isFile()) {
       analyzeName(token, options)
       await analyzeTypes(token, options)
+
+      if (!token.digest) {
+        await computeDigest(token, options)
+      }
+
+      if (options.cache && options.cache.getFileCache && await options.cache.getFileCache(token, options)) {
+        token.unchanged = true
+        token.analyzed = true
+      }
 
       await analyzeArchive(token, options)
       await analyzeFile(token, options)
