+++ lib/url/analyzers/file.js
@@ -19,11 +19,6 @@ async function analyzeBody(token, options) {
     token.fileSize = fileSize
     token.digest = digest
     token.path = filePath
-
-    if (options.cache && options.cache.getFileCache && await options.cache.getFileCache(token, options)) {
-      token.unchanged = true
-      token.analyzed = true
-    }
   } catch (err) {
     token.type = 'file'
     token.warning = err.message
